Move non cmd code into fpgrowth package.
